requirements.txt
python-telegram-bot
persiantools
pytz
